+++ src/fields.js
@@ -2669,6 +2669,194 @@ x3dom.fields.Quaternion = function ( x, y, z, w )
 };
 
 /**
+ * SFRotation constructor.
+ *
+ * Represents a SFRotation, the four components of which, all being
+ * numbers, can be accessed and modified directly.
+ *
+ * Note that the coordinates, if supplied separately or in an array,
+ * are always construed in the order of x, y, z, and w. This fact is
+ * significant as some conventions prepend the scalar component w
+ * instead of listing it as the terminal entity. A quaternion in our
+ * sense, hence, is a quadruple (x, y, z, w).
+ *
+ * @class Represents a SFRotation
+ */
+x3dom.fields.SFRotation = new Proxy( x3dom.fields.Quaternion,
+    {
+        construct : function ( target, args )
+        {
+            args[ 0 ] = args[ 0 ] || 0;
+            args[ 1 ] = args[ 1 ] == undefined ? 1 : args[ 1 ];
+            args[ 2 ] = args[ 2 ] || 0;
+            args[ 3 ] = args[ 3 ] || 0;
+            var quat;
+            var handler = {
+                get : function ( target, prop )
+                {
+                    switch ( prop )
+                    {
+                        case "0":
+                        //case "x":
+                            return target.SFRotation.x;
+                            break;
+                        case "1":
+                        //case "y":
+                            return target.SFRotation.y;
+                            break;
+                        case "2":
+                        //case "z":
+                            return target.SFRotation.z;
+                            break;
+                        case "3":
+                        case "angle":
+                            return target.SFRotation.angle;
+                            break;
+                        default:
+                            return Reflect.get( target, prop );
+                    }
+                },
+                set : function ( target, prop, value )
+                {
+                    var rot = target.SFRotation;
+                    rot[ prop ] = value;
+                    target.setValues(
+                        new x3dom.fields.Quaternion.axisAngle(
+                            new x3dom.fields.SFVec3f( rot.x, rot.y, rot.z ), rot.angle
+                        )
+                    );
+                    return true;
+                }
+            };
+
+            if ( args[ 0 ].constructor == x3dom.fields.SFVec3f )
+            {
+                if ( args[ 1 ].constructor == x3dom.fields.SFVec3f )
+                {
+                    quat = new x3dom.fields.Quaternion.rotateFromTo( args[ 0 ].normalize(), args[ 1 ].normalize() );
+                }
+                else
+                {
+                    quat = new x3dom.fields.Quaternion.axisAngle( args[ 0 ], args[ 1 ] );
+                }
+                //save properties
+                var aa = quat.toAxisAngle();
+                quat.SFRotation = {
+                    x     : aa[ 0 ].x,
+                    y     : aa[ 0 ].y,
+                    z     : aa[ 0 ].z,
+                    angle : aa[ 1 ]
+                };
+            }
+            else
+            {
+                quat = new x3dom.fields.Quaternion.axisAngle( new x3dom.fields.SFVec3f( args[ 0 ], args[ 1 ], args[ 2 ] ), args[ 3 ] );
+                quat.SFRotation = {
+                    x     : args[ 0 ],
+                    y     : args[ 1 ],
+                    z     : args[ 2 ],
+                    angle : args[ 3 ]
+                };
+            }
+            return new Proxy( quat, handler );
+        }
+    } );
+
+/**
+ * SAI function to return the axis of rotation.
+ *
+ * @returns {x3dom.fields.SFVec3f} the axis of rotation
+ */
+x3dom.fields.Quaternion.prototype.getAxis = function ()
+{
+    if ( "SFRotation" in this )
+    {
+        var axis = this.SFRotation;
+        return new x3dom.fields.SFVec3f( axis.x, axis.y, axis.z );
+    }
+    else
+    {
+        var aa = this.toAxisAngle();
+        return aa[ 0 ];
+    }
+};
+
+/**
+ * SAI function to set the axis of rotation.
+ * not well tested
+ * @param {x3dom.fields.SFVec3f} vec - the axis vector to set to.
+ * @returns {} void
+ */
+x3dom.fields.Quaternion.prototype.setAxis = function ( vec )
+{
+    var angle;
+    if ( "SFRotation" in this )
+    {
+        angle = this.SFRotation.angle;
+        this.SFRotation.x = vec.x;
+        this.SFRotation.y = vec.y;
+        this.SFRotation.z = vec.z;
+    }
+    else
+    {
+        angle = this.angle();
+    }
+    var q = new x3dom.fields.Quaternion.axisAngle( vec, angle );
+    this.setValues( q );
+};
+
+/**
+ * SAI function to return the inverse of this object's rotation.
+ * see below
+ * @returns {x3dom.fields.SFRotation} the inverted rotation
+ */
+
+/**
+ * SAI function to return the object multiplied by the passed value.
+ * seee below
+ * @returns {x3dom.fields.SFRotation} the multiplied rotation
+ */
+
+/**
+ * SAI function to return the value of vec multiplied by the matrix corresponding to this object's rotation.
+ *
+ * @param {x3dom.fields.SFVec3f} vec - the vector to multiply with
+ * @returns {x3dom.fields.SFVec3f} the axis of rotation
+ */
+x3dom.fields.Quaternion.prototype.multiVec = function ( vec )
+{
+    var m = x3dom.fields.SFMatrix4f.identity();
+    m.setRotate( this );
+    return m.multMatrixVec( vec );
+};
+
+/**
+ * slerp SAI function to return the value of the spherical linear interpolation between this object's rotation and dest at value 0 ≤ t ≤ 1.
+ * For t = 0, the value is this object`s rotation. For t = 1, the value is the dest rotation.
+ * see below, not well tested
+ * @param {x3dom.fields.SFRotation} dest - the destination rotation
+ * @param {x3dom.fields.SFFloat} t - the fraction
+ * @returns {x3dom.fields.SFRotation} the interpolated rotation
+ */
+
+/**
+ * Sets the components of this quaternion from another quaternion, and returns
+ * this modified quaternion.
+ *
+ * @param {x3dom.fields.Quaternion} that - the quaternion to copy from
+ * @returns {x3dom.fields.Quaternion} this modified quaternion
+ */
+x3dom.fields.Quaternion.prototype.setValues = function ( that )
+{
+    this.x = that.x;
+    this.y = that.y;
+    this.z = that.z;
+    this.w = that.w;
+
+    return this;
+};
+
+/**
  * Returns a copy of the supplied quaternion.
  *
  * @param {x3dom.fields.Quaternion} v - the quatenion to copy
@@ -2690,12 +2878,18 @@ x3dom.fields.Quaternion.copy = function ( v )
  */
 x3dom.fields.Quaternion.prototype.multiply = function ( that )
 {
-    return new x3dom.fields.Quaternion(
+    var product = new x3dom.fields.Quaternion(
         this.w * that.x + this.x * that.w + this.y * that.z - this.z * that.y,
         this.w * that.y + this.y * that.w + this.z * that.x - this.x * that.z,
         this.w * that.z + this.z * that.w + this.x * that.y - this.y * that.x,
         this.w * that.w - this.x * that.x - this.y * that.y - this.z * that.z
     );
+    if ( "SFRotation" in this )
+    {
+        var aa = product.toAxisAngle();
+        return new x3dom.fields.SFRotation( aa[ 0 ].x, aa[ 0 ].y, aa[ 0 ].z, aa[ 1 ] );
+    }
+    return product;
 };
 
 /**
@@ -3128,7 +3322,13 @@ x3dom.fields.Quaternion.prototype.negate = function ()
  */
 x3dom.fields.Quaternion.prototype.inverse = function ()
 {
-    return new x3dom.fields.Quaternion( -this.x, -this.y, -this.z, this.w );
+    var inverse = new x3dom.fields.Quaternion( -this.x, -this.y, -this.z, this.w );
+    if ( "SFRotation" in this )
+    {
+        var aa = inverse.toAxisAngle();
+        return new x3dom.fields.SFRotation( aa[ 0 ].x, aa[ 0 ].y, aa[ 0 ].z, aa[ 1 ] );
+    }
+    return inverse;
 };
 
 /**
@@ -3184,7 +3384,13 @@ x3dom.fields.Quaternion.prototype.slerp = function ( that, t )
     }
 
     // build the new quaternion
-    return this.multScalar( scalerot0 ).add( rot1.multScalar( scalerot1 ) );
+    var result = this.multScalar( scalerot0 ).add( rot1.multScalar( scalerot1 ) );
+    if ( "SFRotation" in this )
+    {
+        var aa = result.toAxisAngle();
+        return new x3dom.fields.SFRotation( aa[ 0 ].x, aa[ 0 ].y, aa[ 0 ].z, aa[ 1 ] );
+    }
+    return result;
 };
 
 /**
@@ -3277,6 +3483,10 @@ x3dom.fields.Quaternion.prototype.toGL = function ()
  */
 x3dom.fields.Quaternion.prototype.toString = function ()
 {
+    if ( "SFRotation" in this )
+    {
+        return this.SFRotation.x + " " + this.SFRotation.y + " " + this.SFRotation.z + " " + this.SFRotation.angle;
+    }
     return this.x + " " + this.y + " " + this.z + ", " + this.w;
 };